* Fix stupid size_t issue in a debug printf

+++ src/String.h
@@ -442,7 +442,7 @@ class String {
         */
         inline bool hasIndex(int i) const { 
 #ifdef DEBUG
-        if (i < 0 || i >= (int) (offset + length())) std::printf("ATTEMPT TO ACCESS INDEX %d/%d\n", i, (offset + length()));
+        if (i < 0 || i >= (int) (offset + length())) std::printf("ATTEMPT TO ACCESS INDEX %d/%zu\n", i, size_t(offset + length()));
 #endif
           return (i < (int) length()); 
         };
